- Fix compile dependency for mekanism - Switch to longs for quantities - Add mekanism chemical resource types - Abstract default impl for resourcetype::collect

+++ src/main/java/ca/teamdman/sfm/common/program/OutputResourceMatcher.java
@@ -3,7 +3,7 @@ package ca.teamdman.sfm.common.program;
 import ca.teamdman.sfml.ast.ResourceLimit;
 
 public class OutputResourceMatcher<STACK, CAP> extends ResourceMatcher<STACK, CAP> {
-    private int seen = 0;
+    private long seen = 0;
 
     public OutputResourceMatcher(ResourceLimit<STACK, CAP> resourceLimit) {
         super(resourceLimit);
@@ -21,18 +21,18 @@ public class OutputResourceMatcher<STACK, CAP> extends ResourceMatcher<STACK, CA
         return false;
     }
 
-    private int getRemainingRoom() {
+    private long getRemainingRoom() {
         return LIMIT.limit().retention() - seen;
     }
 
     @Override
-    public void trackTransfer(int amount) {
+    public void trackTransfer(long amount) {
         super.trackTransfer(amount);
         seen += amount;
     }
 
     @Override
-    public int getMaxTransferable() {
+    public long getMaxTransferable() {
         return Math.min(super.getMaxTransferable(), getRemainingRoom());
     }
 }
