+++ src/main/java/ca/teamdman/sfml/ast/Quantity.java
@@ -1,4 +1,4 @@
 package ca.teamdman.sfml.ast;
 
-public record Quantity(int value) implements ASTNode {
+public record Quantity(long value) implements ASTNode {
 }
